Implement `cumulativeTokens`
This function encapsulates the exponential decay logic controlling total
token minting over time. It's implemented with a Taylor expansion of the
ideal function, truncated at the number of terms necessary to make it
correct within 1 ulp (here, ulp = 1e-6, since we have 1 million tokens).

I've done some preliminary analysis that suggests that it is actually
that accurate, but more tests are warranted, both of the actual EVM code
and of the algorithm. We may want to implement the fixed-point logic in
a faster language so that we can run more exhaustive tests. (It's a bit
awkward, because 256-bit math is actually very convenient here, and Rust
only natively supports 128-bit math, but we can figure something out.)

It looks like `cumulativeTokens` consumes about 145k gas per call (if
the call doesn't return immediately). This feels somewhat high, but also
not the end of the world. Let's wait to optimize until we see in which
cases we actually need to compute this number.