Implement piecewise-linear token vesting
Alternative approach to the existing `cumulativeTokens` method.
Here, instead of computing the exponential curve on the fly, we
initialize the contract with a precomputed list of waypoints and
interpolate linearly between them.

Since the independent variable is time, and we may assume that time
progresses monotonically*, we cache the position into the waypoint
array, so that determining the exact current value only requires zero or
one probes in the common case. Methods that need to update state anyway
(e.g., to claim tokens) will also update the cache as they do so.
(Thanks to @decentralion for suggesting this trick.) One downside is
that it will cost a lot of gas to store these waypoints at contract
initialization time.

Preliminary analysis suggests that with 1000 waypoints, the discrepancy
of the linear curve compared to the exponential curve is less than 1 ppm
for the first 10 years, which seems acceptable. The exponential ideal is
just a guideline, anyway.

* Monotonicity of time is ensured not only by physics but also by the EVM:
each block's `timestamp` must be strictly greater than the last.