Improve exponential decay convergence
This patch makes the on-chain exponential function evaluation more
efficient by computing fewer terms of the Taylor expansion when the
argument is small. At the same time, we improve the precision from
"within `1`" to "within `1e-3`", for parity with the waypoint-based
approach. Even with this improved precision, gas costs are reduced from
170k to 40k by choosing the term count dynamically.

We could improve on this further by also dynamically choosing the Taylor
expansion itself, expanding around a value closer to the input argument
rather than always around 0. The series is still very clean, basically
because `e^z = e^a e^(z - a)` for all `a`, so if you know `e^a` for some
`a` such that `z - a` is close to zero, then you're back within a tight
radius of convergence, and the only penalty is one extra multiplication
by `e^a` (a known constant) at the end.

The added tests pass both before and after this change.