Implement basic token claiming for artists
The contract now tracks an artist address, which can be set by the owner
or transferred by the artist. (The owner is initially the contract
deployer, but ownership is expected to be transferred to the DAO after
contract initialization.)

We still have both the `cumulativeBatches` linear waypoint interpolation
and the `cumulativeTokens` exponential computation. This patch uses the
linear interpolation method for now.

This mechanism is extensible to other claimant types: just add
authorization.