Add on-chain implementation of `log2`
This uses the fast binary logarithm given by Clay S. Turner:
<http://www.claysturner.com/dsp/BinaryLogarithm.pdf>

It ports very nicely to the EVM. We emit the full 64 mantissa bits of
precision, though in practice we may need well less than that, which we
can hone later.

This is needed to compute the inverse function of `cumulativeTokens`:
i.e., to determine the time after which `n` tokens have fully vested.