Update caching and claiming to use `batchesVestedInverse`
This new approach takes advantages of both the stored waypoints and the
exact exponential function. We still use linear interpolation, so we
have continuous vesting. We still use an on-chain cache that only needs
to be updated on new batches, so computation costs are low. But now,
instead of pre-populating the entirety of the cache at contract
initialization time, we compute cache members on chain and on demand.

Specifically: we cache the number of fully vested batches `n`, and we
cache relative timestamps `lo = t(n)` and `hi = t(n + 1)`, where `t(m)`
gives the relative timestamp by which exactly `m` batches have fully
vested. Then, in the usual case, `block.timestamp` falls between `lo`
and `hi`, and so only linear interpolation is needed. Once the timestamp
exceeds `hi`, we linearly probe `n` until `t(n + 1) > block.timestamp`,
recompute `lo` and `hi`, and update the cache.

Compared to the stored waypoints approach, this approach can use
arbitrarily precise buckets for linearization with no extra storage
cost. (The cost of increasing the precision is that the cache has to be
updated more frequently, so a higher proportion of claims will incur the
modest expense of re-computing the logarithm and storing to the cache.)