Use 30-bit logarithm for token vesting
Reduces from 64- to 30-bit precision. The `batchVestingCurve.js` script
shows that this causes 53 (out of 1001) entries to be off by 1 second, a
relative error on the order of under 1 part per million.

This seems acceptable, given that it also improves that script's runtime
from 68 seconds to 33 seconds on my laptop (i5-4300U) and reduces the
cost of a logarithm from 113k gas to 63k gas.