Implement token claiming for bots
The core claiming logic uses the same codepaths as artists. What's new
is the authorization checking: you can claim tokens on behalf of a bot
if you (a) hold its NFT, (b) are approved as an operator of that NFT,
or (c) are approved as a universal operator for the holder of the NFT.

For efficiency, we hard-code the mapping between `botId` and `nftId`:
i.e., the map `(x) => 40000000 + x`. This seems stable enough, and in
the event that ArtBlocks upgrades its contract and changes around all
the IDs (seems unlikely), we can re-point our ArtBlocks contract at a
wrapper `IERC721` implementer that translates ID formats and calls out
to the real ArtBlocks.