Always refuse to claim tokens to the zero address
For the community and treasury splits, refusing to send to the zero
address is a security feature, so that arbitrary users can't burn these
communal tokens before the community and treasury addresses are
initialized. For bots and artists, an authorized caller has control over
the destination and so can always burn the tokens manually, but refusing
to send directly to the zero address at claim time still seems like a
nice safety check on behalf of the caller.