Exhaustively test `batchesVestedInverse`
This test is slow: it takes 8.5 seconds on my laptop (i5-4300U).
But it's pretty bulletproof, testing a declared-`pure` function on its
entire domain, and since this covers the entirety of the fixed-point
math, it seems plausibly worth it.

To verify visually that this represents a reasonable curve:

```
jq .[] test/expectedVestingCurve.json | gnuplot --persist -e 'plot "-"'
```